+++ src/test/java/lotto/dto/WonTest.java
@@ -1,6 +1,6 @@
 package lotto.dto;
 
-import lotto.model.Won;
+import lotto.Won;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
